Create 0x-swap-utils file
