+++ background/redux-slices/0x-swap.ts
@@ -28,16 +28,13 @@ import {
 } from "../constants"
 import { EVMNetwork } from "../networks"
 import { setSnackbarMessage } from "./ui"
+import { enrichAssetAmountWithDecimalValues } from "./utils/asset-utils"
+import { AssetsState } from "./assets"
 import {
-  enrichAssetAmountWithDecimalValues,
-  enrichAssetAmountWithMainCurrencyValues,
-} from "./utils/asset-utils"
-import {
-  convertFixedPointNumber,
-  fixedPointNumberToString,
-  parseToFixedPointNumber,
-} from "../lib/fixed-point"
-import { AssetsState, selectAssetPricePoint } from "./assets"
+  calculatePriceDetails,
+  PriceDetails,
+  SwapQuoteRequest,
+} from "./utils/0x-swap-utils"
 
 // This is how 0x represents native token addresses
 const ZEROEX_NATIVE_TOKEN_CONTRACT_ADDRESS =
@@ -56,35 +53,8 @@ type ZeroExValidationError = {
   reason: string
 }
 
-interface SwapAssets {
-  sellAsset: SmartContractFungibleAsset | FungibleAsset
-  buyAsset: SmartContractFungibleAsset | FungibleAsset
-}
-
-type SwapAmount =
-  | {
-      sellAmount: string
-    }
-  | {
-      buyAmount: string
-    }
-
-export type SwapQuoteRequest = {
-  assets: SwapAssets
-  amount: SwapAmount
-  slippageTolerance: number
-  gasPrice: bigint
-  network: EVMNetwork
-}
-
 export type ZrxPrice = ValidatedType<typeof isValidSwapPriceResponse>
 export type ZrxQuote = ValidatedType<typeof isValidSwapQuoteResponse>
-
-export type PriceDetails = {
-  priceImpact: number | undefined
-  buyCurrencyAmount: string | undefined
-  sellCurrencyAmount: string | undefined
-}
 
 export interface SwapState {
   latestQuoteRequest?: SwapQuoteRequest | undefined
@@ -337,86 +307,6 @@ const parseAndNotifyOnZeroExApiError = (
   }
 }
 
-const getAssetAmount = (
-  assets: AssetsState,
-  asset: SmartContractFungibleAsset | FungibleAsset,
-  amount: string
-) => {
-  const mainCurrencySymbol = "USD"
-
-  const fixedPointAmount = parseToFixedPointNumber(amount.toString())
-  if (typeof fixedPointAmount === "undefined") {
-    return undefined
-  }
-  const decimalMatched = convertFixedPointNumber(
-    fixedPointAmount,
-    asset.decimals
-  )
-
-  const assetPricePoint = selectAssetPricePoint(
-    assets,
-    asset?.symbol,
-    mainCurrencySymbol
-  )
-  return enrichAssetAmountWithMainCurrencyValues(
-    {
-      asset,
-      amount: decimalMatched.amount,
-    },
-    assetPricePoint,
-    2
-  )
-}
-
-const getPriceImpact = (
-  buyCurrencyAmount: number | undefined,
-  sellCurrencyAmount: number | undefined
-): number | undefined => {
-  if (buyCurrencyAmount && sellCurrencyAmount) {
-    return +(buyCurrencyAmount / sellCurrencyAmount - 1).toFixed(2)
-  }
-  return undefined
-}
-
-const calculatePriceDetails = (
-  quoteRequest: SwapQuoteRequest,
-  assets: AssetsState,
-  sellAmount: string,
-  buyAmount: string,
-  dispatch: ThunkDispatch<unknown, unknown, AnyAction>
-) => {
-  const assetSellAmount = getAssetAmount(
-    assets,
-    quoteRequest.assets.sellAsset,
-    fixedPointNumberToString({
-      amount: BigInt(sellAmount),
-      decimals: quoteRequest.assets.sellAsset.decimals,
-    })
-  )
-
-  const assetBuyAmount = getAssetAmount(
-    assets,
-    quoteRequest.assets.buyAsset,
-    fixedPointNumberToString({
-      amount: BigInt(buyAmount),
-      decimals: quoteRequest.assets.buyAsset.decimals,
-    })
-  )
-
-  const priceImpact = getPriceImpact(
-    assetBuyAmount?.mainCurrencyAmount,
-    assetSellAmount?.mainCurrencyAmount
-  )
-
-  const data: PriceDetails = {
-    buyCurrencyAmount: assetBuyAmount?.localizedMainCurrencyAmount,
-    sellCurrencyAmount: assetSellAmount?.localizedMainCurrencyAmount,
-    priceImpact,
-  }
-
-  dispatch(setPriceDetails(data))
-}
-
 /**
  * This async thunk fetches an indicative RFQ-T price for a swap. The quote
  * request specifies the swap assets as well as one end of the swap, and the
@@ -484,13 +374,14 @@ export const fetchSwapPrice = createBackgroundAsyncThunk(
 
       dispatch(setLatestQuoteRequest(quoteRequest))
 
-      calculatePriceDetails(
+      const priceDetails = calculatePriceDetails(
         quoteRequest,
         assets,
         quote.sellAmount,
-        quote.buyAmount,
-        dispatch
+        quote.buyAmount
       )
+
+      dispatch(setPriceDetails(priceDetails))
 
       return { quote, needsApproval }
     } catch (error) {
