+++ ui/pages/Swap.tsx
@@ -11,7 +11,6 @@ import {
   approveTransfer,
   selectLatestQuoteRequest,
   selectInProgressApprovalContract,
-  SwapQuoteRequest,
   fetchSwapQuote,
   fetchSwapPrice,
   selectPriceDetails,
@@ -45,6 +44,7 @@ import { selectSlippageTolerance } from "@tallyho/tally-background/redux-slices/
 import { isNetworkBaseAsset } from "@tallyho/tally-background/redux-slices/utils/asset-utils"
 import { ReadOnlyAccountSigner } from "@tallyho/tally-background/services/signing"
 import { EIP_1559_COMPLIANT_CHAIN_IDS } from "@tallyho/tally-background/constants"
+import { SwapQuoteRequest } from "@tallyho/tally-background/redux-slices/utils/0x-swap-utils"
 import CorePage from "../components/Core/CorePage"
 import SharedAssetInput from "../components/Shared/SharedAssetInput"
 import SharedButton from "../components/Shared/SharedButton"
